atelet: simplify the file cache internals per review
Remove the createDestFile test seam and the sparseDest indirection:
CopyFile creates its destination directly, copySparse writes real
files, and the userspace fallback is tested through the public API by
copying across filesystems (copy_file_range fails with EXDEV on a
/dev/shm destination). Name getTo's serving-mode parameter (serveHit)
so the shared loop reads as the strategy split it is, not a test seam.

Eviction passes now retry .rm-* leftovers a previous pass failed to
remove, so those bytes no longer wait for the next restart's sweep;
tmp/ stays startup-swept only, since a later sweep would delete
in-flight fetches' working directories.

Split filecache.go into doc.go, store.go, and sweep.go, moving the
sizing helpers next to their consumer in evict.go, and shorten the
package's comments.
