--- cmd/atelet/internal/filecache/filecache.go
@@ -1,298 +1,0 @@
-// Copyright 2026 Google LLC
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-// Package filecache is a node-local disk cache of downloaded artifacts,
-// keyed by opaque identities (see Key). Its contract: an artifact wanted by
-// N concurrent callers is fetched once, publication into the cache is atomic
-// and crash-safe, and cached bytes are evicted under byte-budget pressure
-// without ever breaking a consumer.
-//
-// On-disk layout, under a store's root (which the store owns exclusively):
-//
-//	entries/<sha256(key)>/
-//	  data       # the cached file (or directory tree)
-//	  meta.json  # canonical key + creation time; debugging only
-//	tmp/         # in-flight fetches; same filesystem as entries/, so
-//	             # publication is one atomic rename
-//	.rm-*        # retired entries awaiting removal
-//
-// An entry directory's mtime is its last-use clock. Nothing on a correctness
-// path reads meta.json: entries are matched to keys by hashing the keys.
-//
-// A crash can leave debris in tmp/ (a fetch that never finished) or .rm-*
-// dirs (an eviction that renamed but never removed); SweepDebris reaps both
-// and runs once at startup, before the store serves requests. Everything
-// under entries/ is a complete, published entry.
-//
-// # Contracts
-//
-// Consumers are protected from eviction by construction, never by luck:
-//
-//   - GetFileTo hard-links the cached file to the consumer's path in the
-//     same locked step that resolves the hit, so once it returns, the
-//     consumer's file is valid forever — evicting the entry later only
-//     removes the cache's own link. Eviction can cost the next caller a
-//     refetch, never break a served one.
-//   - GetFileCopyTo hands the consumer a private copy instead: the cached
-//     inode is never shared, so the consumer may mutate the result, and a
-//     copy in flight reads a held-open handle whose bytes outlive any
-//     concurrent eviction.
-//   - The store's min age (WithMinAge) vetoes eviction of entries younger
-//     than the worst-case window between publication and a consumer's use
-//     becoming visible; size it for the slowest consumer.
-//
-// Cached bytes are shared through links: entries are published read-only
-// (0444) so a consumer that tries to write through its hard link fails with
-// EACCES instead of corrupting the copy every later caller receives.
-// Callers that must mutate an artifact in place use GetFileCopyTo.
-//
-// Keys are identities, not addresses. The store never invalidates: a key
-// must name content that is immutable at its source (see URIKey), and a
-// changed artifact must arrive under a new key.
-package filecache
-
-import (
-	"context"
-	"encoding/json"
-	"errors"
-	"fmt"
-	"io/fs"
-	"os"
-	"path/filepath"
-	"strings"
-	"sync"
-	"syscall"
-	"time"
-
-	"golang.org/x/sync/singleflight"
-)
-
-const (
-	entriesDirName = "entries"
-	tmpDirName     = "tmp"
-	// rmPrefix marks a retired entry awaiting removal, at the store root (not
-	// under entries/, so a retired entry is invisible to lookups and GC
-	// listings the moment it is renamed).
-	rmPrefix = ".rm-"
-
-	dataName = "data"
-	metaName = "meta.json"
-
-	// defaultMinAge is the default eviction minimum age (see WithMinAge).
-	defaultMinAge = 10 * time.Minute
-	// defaultFetchTimeout is the default per-fetch bound (see
-	// WithFetchTimeout). Generous enough for multi-GiB artifacts on a busy
-	// node.
-	defaultFetchTimeout = 10 * time.Minute
-)
-
-// Store is one on-disk cache. It is safe for concurrent use and assumes it
-// is the only writer under its root (one atelet per node).
-type Store struct {
-	root string
-
-	// minAge vetoes eviction of any entry younger than this, covering the
-	// window between publication and the consumer's use becoming visible to
-	// GC (a hardlink's Nlink, or a root-set record).
-	minAge time.Duration
-
-	// fetchTimeout bounds each fetch. Fetches run detached from the contexts
-	// of the callers waiting on them, so this is the only bound on how long
-	// one can run.
-	fetchTimeout time.Duration
-
-	// sf collapses concurrent fetches of the same key into one flight.
-	// Eviction will retire entries inside the same flight, so a retire can
-	// never race a fetch of the key it is removing.
-	sf singleflight.Group
-
-	// hitMu closes the hit-vs-evict window: held shared by the link-out path
-	// (stat, link, and last-use touch), exclusive by eviction's final
-	// re-check and retire rename, so an entry can never vanish between a
-	// hit's stat and its link. Uncontended except during an eviction pass.
-	hitMu sync.RWMutex
-
-	// evictMu serializes EvictUnused passes (concurrent passes would fight
-	// over the same candidates for no benefit).
-	evictMu sync.Mutex
-}
-
-// Option configures a Store.
-type Option func(*Store)
-
-// WithMinAge sets the eviction minimum age.
-func WithMinAge(d time.Duration) Option {
-	return func(s *Store) { s.minAge = d }
-}
-
-// WithFetchTimeout sets the per-fetch bound.
-func WithFetchTimeout(d time.Duration) Option {
-	return func(s *Store) { s.fetchTimeout = d }
-}
-
-// New opens (creating if needed) the store rooted at root.
-func New(root string, opts ...Option) (*Store, error) {
-	s := &Store{
-		root:         root,
-		minAge:       defaultMinAge,
-		fetchTimeout: defaultFetchTimeout,
-	}
-	for _, opt := range opts {
-		opt(s)
-	}
-	// Entries are world-readable (their files get hard-linked into consumer
-	// dirs and, later, consumed in place); tmp holds unpublished fetches and
-	// stays private.
-	if err := os.MkdirAll(s.entriesDir(), 0o755); err != nil {
-		return nil, fmt.Errorf("while creating entries dir: %w", err)
-	}
-	if err := os.MkdirAll(s.tmpDir(), 0o700); err != nil {
-		return nil, fmt.Errorf("while creating tmp dir: %w", err)
-	}
-	return s, nil
-}
-
-func (s *Store) entriesDir() string { return filepath.Join(s.root, entriesDirName) }
-func (s *Store) tmpDir() string     { return filepath.Join(s.root, tmpDirName) }
-
-// entryDir is the published location of k's entry.
-func (s *Store) entryDir(k Key) string { return filepath.Join(s.entriesDir(), k.dir) }
-
-// dataPath is the published location of k's cached file (or tree).
-func (s *Store) dataPath(k Key) string { return filepath.Join(s.entryDir(k), dataName) }
-
-// entryMeta is the debugging sidecar written next to an entry's data. It is
-// never read on a correctness path; a missing or corrupt one affects
-// nothing.
-type entryMeta struct {
-	// Key is the canonical key string, so an operator staring at du output
-	// can tell what an entry holds.
-	Key       string    `json:"key"`
-	CreatedAt time.Time `json:"createdAt"`
-}
-
-// writeEntryMeta writes the meta.json sidecar into an (unpublished) entry
-// dir.
-func writeEntryMeta(entryDir string, k Key, createdAt time.Time) error {
-	data, err := json.Marshal(entryMeta{Key: k.String(), CreatedAt: createdAt})
-	if err != nil {
-		return fmt.Errorf("while marshaling entry meta: %w", err)
-	}
-	if err := os.WriteFile(filepath.Join(entryDir, metaName), data, 0o644); err != nil {
-		return fmt.Errorf("while writing entry meta: %w", err)
-	}
-	return nil
-}
-
-// SweepStats reports what a SweepDebris pass removed.
-type SweepStats struct {
-	// TmpRemoved counts removed tmp/ children (unfinished fetches).
-	TmpRemoved int
-	// RetiredRemoved counts removed .rm-* dirs (interrupted evictions).
-	RetiredRemoved int
-}
-
-// SweepDebris removes crash debris: everything under tmp/ and every .rm-*
-// dir at the store root. It runs once at startup, before the store serves
-// requests; published entries are never touched. Removal failures are
-// joined and reported after the sweep visits everything, so one bad path
-// does not shadow the rest.
-func (s *Store) SweepDebris(ctx context.Context) (SweepStats, error) {
-	var stats SweepStats
-	var errs []error
-
-	tmpChildren, err := os.ReadDir(s.tmpDir())
-	if err != nil {
-		errs = append(errs, fmt.Errorf("while listing tmp dir: %w", err))
-	}
-	for _, child := range tmpChildren {
-		if err := ctx.Err(); err != nil {
-			return stats, err
-		}
-		if err := os.RemoveAll(filepath.Join(s.tmpDir(), child.Name())); err != nil {
-			errs = append(errs, fmt.Errorf("while removing tmp debris %q: %w", child.Name(), err))
-			continue
-		}
-		stats.TmpRemoved++
-	}
-
-	rootChildren, err := os.ReadDir(s.root)
-	if err != nil {
-		errs = append(errs, fmt.Errorf("while listing store root: %w", err))
-	}
-	for _, child := range rootChildren {
-		if err := ctx.Err(); err != nil {
-			return stats, err
-		}
-		if !strings.HasPrefix(child.Name(), rmPrefix) {
-			continue
-		}
-		if err := os.RemoveAll(filepath.Join(s.root, child.Name())); err != nil {
-			errs = append(errs, fmt.Errorf("while removing retired entry %q: %w", child.Name(), err))
-			continue
-		}
-		stats.RetiredRemoved++
-	}
-
-	return stats, errors.Join(errs...)
-}
-
-// allocatedSize returns the bytes a file actually occupies on disk
-// (st_blocks), not its logical length. The distinction matters here more
-// than almost anywhere: the cache's biggest artifacts are sparse guest
-// memory images, whose logical size can exceed their footprint by orders of
-// magnitude, and every byte figure the store reports is compared against
-// real disk usage (statfs watermarks, byte budgets) — logical sizes would
-// trigger eviction that frees nothing and report targets met that were not.
-func allocatedSize(info fs.FileInfo) int64 {
-	if st, ok := info.Sys().(*syscall.Stat_t); ok {
-		return st.Blocks * 512
-	}
-	return info.Size()
-}
-
-// TotalBytes sums the allocated (on-disk) bytes of all published entries'
-// regular files. It is the GC driver's usage measure against the store's
-// byte budget.
-func (s *Store) TotalBytes(ctx context.Context) (int64, error) {
-	var total int64
-	err := filepath.WalkDir(s.entriesDir(), func(path string, d fs.DirEntry, err error) error {
-		if err != nil {
-			// An entry retired mid-walk is not an error; skip what vanished.
-			if errors.Is(err, fs.ErrNotExist) {
-				return nil
-			}
-			return err
-		}
-		if err := ctx.Err(); err != nil {
-			return err
-		}
-		if !d.Type().IsRegular() {
-			return nil
-		}
-		info, err := d.Info()
-		if err != nil {
-			if errors.Is(err, fs.ErrNotExist) {
-				return nil
-			}
-			return err
-		}
-		total += allocatedSize(info)
-		return nil
-	})
-	if err != nil {
-		return 0, fmt.Errorf("while sizing entries: %w", err)
-	}
-	return total, nil
-}
